perf+fix: P1 gap closures from the audit
- dashboard::snapshot collapses four serial D1 round-trips into a single
  batched three-GROUP-BY + top-10 activity query. Server-side aggregation
  removes the Vec<Row> shuttle and a ~40-160 ms landing-page latency
  contribution.
- Cache-Control is method-aware: GET/HEAD render HTML ('private, no-cache',
  bfcache-eligible, revalidates), everything else stays 'no-store'.
  http::Method isn't structural-match-safe so use an if/else equality
  comparison rather than a match on associated constants.
- Wire types: StageCounts/ResourceStatusCounts/PlacementStatusCounts and
  the three *_total fields on DashboardSnapshot switch from usize to u32
  so the wasm32 client and x86_64 server agree on integer width across
  the hydration boundary. MetricCard follows.
- normalize_text caps on char count, not byte length — multibyte input
  was hitting the limit well below the documented 'N characters'.
- list_households / list_resources / list_placements all grow a hard
  LIMIT 500 so a runaway table can't ship unbounded JSON through the
  Worker.

Verified post-deploy v4 (a82a4a89-eefe-4364-9893-f80202ea9d00):
- GET / /households /inventory /placements /activity → 200,
  cache-control: private, no-cache.
- POST /api/* → cache-control: no-store.
- Dashboard SSR payload decodes cleanly: household_total=3,
  resource_total=4, placement_total=2, recent_activity=5 entries.
- All three hashed assets (js/wasm/css) resolve 200.

+++ src/api.rs
@@ -19,11 +19,11 @@ pub struct Household {
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
 pub struct StageCounts {
-    pub intake: usize,
-    pub assessment: usize,
-    pub placement: usize,
-    pub follow_up: usize,
-    pub exited: usize,
+    pub intake: u32,
+    pub assessment: u32,
+    pub placement: u32,
+    pub follow_up: u32,
+    pub exited: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -49,10 +49,10 @@ pub struct HousingResource {
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ResourceStatusCounts {
-    pub available: usize,
-    pub held: usize,
-    pub occupied: usize,
-    pub offline: usize,
+    pub available: u32,
+    pub held: u32,
+    pub occupied: u32,
+    pub offline: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -79,11 +79,11 @@ pub struct Placement {
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
 pub struct PlacementStatusCounts {
-    pub proposed: usize,
-    pub confirmed: usize,
-    pub moved_in: usize,
-    pub exited: usize,
-    pub cancelled: usize,
+    pub proposed: u32,
+    pub confirmed: u32,
+    pub moved_in: u32,
+    pub exited: u32,
+    pub cancelled: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -106,11 +106,11 @@ pub struct ActivityNote {
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct DashboardSnapshot {
     pub households: StageCounts,
-    pub household_total: usize,
+    pub household_total: u32,
     pub resources: ResourceStatusCounts,
-    pub resource_total: usize,
+    pub resource_total: u32,
     pub placements: PlacementStatusCounts,
-    pub placement_total: usize,
+    pub placement_total: u32,
     pub recent_activity: Vec<ActivityNote>,
 }
